reading from a pipe might cause partial buffer reads
reading from a pipe might cause partial buffer reads. When we write we always
want to write a complete buffer, otherwise writing with O_DIRECT fails.
The loop around the wright is probably necessary if writing to a pipe is
going to block and it returns before it blocks for the first time.