detect incomplete streams
We had no way to detect an incomplete stream on the receiving side so far,
if the input was truncated in the middle of some data chunk
(fixed by the previous commit),
or truncated at some chunk boundary,
or just empty.

Add new "chunk types" as BEGIN/END stream markers.

If the stream starts with a BEGIN marker,
a missing END marker will cause a failure.

As backward compatibility, allow for them to be missing,
but complain about it on stderr.